feat: rename GameStatus component to BoardStatus
Move play status texts to this component
