+++ app/ui/game/board-status.module.css
